--- frontend/src/types/RepoFile.ts
@@ -1,4 +1,0 @@
-export interface RepoFile {
-    filename: string,
-    content: string
-}
